requirements.txt
pandas
numpy
fastapi==0.110.0
uvicorn==0.29.0
langchain
chromadb
pypdf
tiktoken
openai
python-dotenv
langchain-community 
langchain_openai
langchain_pinecone
pinecone
gradio==4.44.0
huggingface_hub==0.23.5
pydantic
httpx
requests